refactor: Run IntelliJ cleanup

+++ src/main/java/net/minecraft/block/BlockDispenser.java
@@ -1,14 +1,9 @@
 package net.minecraft.block;
 
-import java.util.Random;
 import net.minecraft.block.material.Material;
 import net.minecraft.client.renderer.texture.IIconRegister;
 import net.minecraft.creativetab.CreativeTabs;
-import net.minecraft.dispenser.BehaviorDefaultDispenseItem;
-import net.minecraft.dispenser.IBehaviorDispenseItem;
-import net.minecraft.dispenser.IBlockSource;
-import net.minecraft.dispenser.IPosition;
-import net.minecraft.dispenser.PositionImpl;
+import net.minecraft.dispenser.*;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.item.EntityItem;
 import net.minecraft.entity.player.EntityPlayer;
@@ -23,6 +18,8 @@ import net.minecraft.util.IIcon;
 import net.minecraft.util.IRegistry;
 import net.minecraft.util.RegistryDefaulted;
 import net.minecraft.world.World;
+
+import java.util.Random;
 
 public class BlockDispenser extends BlockContainer
 {
@@ -109,10 +106,9 @@ public class BlockDispenser extends BlockContainer
         if (p_149727_1_.isClient)
         {
             return true;
-        }
-        else
+        } else
         {
-            TileEntityDispenser var10 = (TileEntityDispenser)p_149727_1_.getTileEntity(p_149727_2_, p_149727_3_, p_149727_4_);
+            TileEntityDispenser var10 = (TileEntityDispenser) p_149727_1_.getTileEntity(p_149727_2_, p_149727_3_, p_149727_4_);
 
             if (var10 != null)
             {
@@ -126,7 +122,7 @@ public class BlockDispenser extends BlockContainer
     protected void func_149941_e(World p_149941_1_, int p_149941_2_, int p_149941_3_, int p_149941_4_)
     {
         BlockSourceImpl var5 = new BlockSourceImpl(p_149941_1_, p_149941_2_, p_149941_3_, p_149941_4_);
-        TileEntityDispenser var6 = (TileEntityDispenser)var5.getBlockTileEntity();
+        TileEntityDispenser var6 = (TileEntityDispenser) var5.getBlockTileEntity();
 
         if (var6 != null)
         {
@@ -135,8 +131,7 @@ public class BlockDispenser extends BlockContainer
             if (var7 < 0)
             {
                 p_149941_1_.playAuxSFX(1001, p_149941_2_, p_149941_3_, p_149941_4_, 0);
-            }
-            else
+            } else
             {
                 ItemStack var8 = var6.getStackInSlot(var7);
                 IBehaviorDispenseItem var9 = this.func_149940_a(var8);
@@ -152,7 +147,7 @@ public class BlockDispenser extends BlockContainer
 
     protected IBehaviorDispenseItem func_149940_a(ItemStack p_149940_1_)
     {
-        return (IBehaviorDispenseItem)dispenseBehaviorRegistry.getObject(p_149940_1_.getItem());
+        return (IBehaviorDispenseItem) dispenseBehaviorRegistry.getObject(p_149940_1_.getItem());
     }
 
     public void onNeighborBlockChange(World p_149695_1_, int p_149695_2_, int p_149695_3_, int p_149695_4_, Block p_149695_5_)
@@ -165,8 +160,7 @@ public class BlockDispenser extends BlockContainer
         {
             p_149695_1_.scheduleBlockUpdate(p_149695_2_, p_149695_3_, p_149695_4_, this, this.tickRate(p_149695_1_));
             p_149695_1_.setBlockMetadataWithNotify(p_149695_2_, p_149695_3_, p_149695_4_, var7 | 8, 4);
-        }
-        else if (!var6 && var8)
+        } else if (!var6 && var8)
         {
             p_149695_1_.setBlockMetadataWithNotify(p_149695_2_, p_149695_3_, p_149695_4_, var7 & -9, 4);
         }
@@ -201,13 +195,13 @@ public class BlockDispenser extends BlockContainer
 
         if (p_149689_6_.hasDisplayName())
         {
-            ((TileEntityDispenser)p_149689_1_.getTileEntity(p_149689_2_, p_149689_3_, p_149689_4_)).func_146018_a(p_149689_6_.getDisplayName());
+            ((TileEntityDispenser) p_149689_1_.getTileEntity(p_149689_2_, p_149689_3_, p_149689_4_)).func_146018_a(p_149689_6_.getDisplayName());
         }
     }
 
     public void breakBlock(World p_149749_1_, int p_149749_2_, int p_149749_3_, int p_149749_4_, Block p_149749_5_, int p_149749_6_)
     {
-        TileEntityDispenser var7 = (TileEntityDispenser)p_149749_1_.getTileEntity(p_149749_2_, p_149749_3_, p_149749_4_);
+        TileEntityDispenser var7 = (TileEntityDispenser) p_149749_1_.getTileEntity(p_149749_2_, p_149749_3_, p_149749_4_);
 
         if (var7 != null)
         {
@@ -231,17 +225,17 @@ public class BlockDispenser extends BlockContainer
                         }
 
                         var9.stackSize -= var13;
-                        EntityItem var14 = new EntityItem(p_149749_1_, (double)((float)p_149749_2_ + var10), (double)((float)p_149749_3_ + var11), (double)((float)p_149749_4_ + var12), new ItemStack(var9.getItem(), var13, var9.getItemDamage()));
+                        EntityItem var14 = new EntityItem(p_149749_1_, (float) p_149749_2_ + var10, (float) p_149749_3_ + var11, (float) p_149749_4_ + var12, new ItemStack(var9.getItem(), var13, var9.getItemDamage()));
 
                         if (var9.hasTagCompound())
                         {
-                            var14.getEntityItem().setTagCompound((NBTTagCompound)var9.getTagCompound().copy());
+                            var14.getEntityItem().setTagCompound((NBTTagCompound) var9.getTagCompound().copy());
                         }
 
                         float var15 = 0.05F;
-                        var14.motionX = (double)((float)this.field_149942_b.nextGaussian() * var15);
-                        var14.motionY = (double)((float)this.field_149942_b.nextGaussian() * var15 + 0.2F);
-                        var14.motionZ = (double)((float)this.field_149942_b.nextGaussian() * var15);
+                        var14.motionX = (float) this.field_149942_b.nextGaussian() * var15;
+                        var14.motionY = (float) this.field_149942_b.nextGaussian() * var15 + 0.2F;
+                        var14.motionZ = (float) this.field_149942_b.nextGaussian() * var15;
                         p_149749_1_.spawnEntityInWorld(var14);
                     }
                 }
@@ -256,9 +250,9 @@ public class BlockDispenser extends BlockContainer
     public static IPosition func_149939_a(IBlockSource p_149939_0_)
     {
         EnumFacing var1 = func_149937_b(p_149939_0_.getBlockMetadata());
-        double var2 = p_149939_0_.getX() + 0.7D * (double)var1.getFrontOffsetX();
-        double var4 = p_149939_0_.getY() + 0.7D * (double)var1.getFrontOffsetY();
-        double var6 = p_149939_0_.getZ() + 0.7D * (double)var1.getFrontOffsetZ();
+        double var2 = p_149939_0_.getX() + 0.7D * (double) var1.getFrontOffsetX();
+        double var4 = p_149939_0_.getY() + 0.7D * (double) var1.getFrontOffsetY();
+        double var6 = p_149939_0_.getZ() + 0.7D * (double) var1.getFrontOffsetZ();
         return new PositionImpl(var2, var4, var6);
     }
 
@@ -274,6 +268,6 @@ public class BlockDispenser extends BlockContainer
 
     public int getComparatorInputOverride(World p_149736_1_, int p_149736_2_, int p_149736_3_, int p_149736_4_, int p_149736_5_)
     {
-        return Container.calcRedstoneFromInventory((IInventory)p_149736_1_.getTileEntity(p_149736_2_, p_149736_3_, p_149736_4_));
+        return Container.calcRedstoneFromInventory((IInventory) p_149736_1_.getTileEntity(p_149736_2_, p_149736_3_, p_149736_4_));
     }
 }
