+++ src/main/java/net/minecraft/block/BlockSoulSand.java
@@ -23,7 +23,7 @@ public class BlockSoulSand extends Block
     public AxisAlignedBB getCollisionBoundingBoxFromPool(World p_149668_1_, int p_149668_2_, int p_149668_3_, int p_149668_4_)
     {
         float var5 = 0.125F;
-        return AxisAlignedBB.getAABBPool().getAABB((double)p_149668_2_, (double)p_149668_3_, (double)p_149668_4_, (double)(p_149668_2_ + 1), (double)((float)(p_149668_3_ + 1) - var5), (double)(p_149668_4_ + 1));
+        return AxisAlignedBB.getAABBPool().getAABB(p_149668_2_, p_149668_3_, p_149668_4_, p_149668_2_ + 1, (float) (p_149668_3_ + 1) - var5, p_149668_4_ + 1);
     }
 
     public void onEntityCollidedWithBlock(World p_149670_1_, int p_149670_2_, int p_149670_3_, int p_149670_4_, Entity p_149670_5_)
